# perf: reduce NewCarReader allocations